fix: move the alfred package up in the hierarchy out of the src/ folder. * move files. * fix import references. * update pyproject.toml file.
