+++ alfred/__init__.py
@@ -1,0 +1,2 @@
+
+__version__ = "0.4.1"
