+++ alfred/base/constants.py
@@ -1,6 +1,6 @@
 from enum import Enum
 
-from src.alfred.http.typed import ResponseType
+from alfred.http.typed import ResponseType
 
 # Response type/header mapping
 RESPONSE_TYPE_HEADER_MAPPING = {
