+++ alfred/base/exceptions.py
