+++ alfred/rest/jobs/typed.py
