+++ alfred/typings/__init__.py
@@ -1,0 +1,1 @@
+from alfred.typings.misc import *
