+++ alfred/typings/misc.py
