--- src/alfred/typings/__init__.py
@@ -1,1 +1,0 @@
-from src.alfred.typings.misc import *
