--- src/alfred/utils/__init__.py
@@ -1,1 +1,0 @@
-from src.alfred.utils.logging import *
